second commit with some code modification and codes written in class controller, repository, service, etc.

+++ src/main/java/com/javams/paymentservice/repository/PaymentRepository.java
@@ -1,10 +1,10 @@
 package com.javams.paymentservice.repository;
 
-import com.javams.paymentservice.entity.Payment;
+import com.javams.paymentservice.model.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, Integer> {
+public interface PaymentRepository extends JpaRepository<Payment, Long> {
 }
